rename manage_my_trends and user_reel packages
