+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/screen/manage_my_trends/mapper.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.presentation.screen.managemytrendscreen
+package net.thechance.mena.trends.presentation.screen.manage_my_trends
 
 import net.thechance.mena.trends.domain.entity.Reel
