+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/screen/user_reel/UserReelEffect.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.presentation.screen.mange_my_trends
+package net.thechance.mena.trends.presentation.screen.user_reel
 
 sealed interface UserReelEffect {
     data object NavigateBack : UserReelEffect
