+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/screen/user_reel/UserReelInteractionListener.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.trends.presentation.screen.mange_my_trends
+package net.thechance.mena.trends.presentation.screen.user_reel
 
 interface UserReelInteractionListener {
     fun onBackClick()
